requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
PyMySQL==1.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
sentence-transformers
google-generativeai
pandas>=2.2.0
